Add config options to configure networks and sysctls
"networks" is a comma-separated list of docker networks that the
container will be connected to once it has been created.

"sysctls" is a comma-separated list of sysctls to apply to the
kernel within the container, e.g.

    sysctls=net.ipv4.conf.all.rp_filter=0,net.ipv4.ip_forward=0

Both new options can be applied to a specific user or to the
DEFAULT section.